move trainer pages to /trainer
