+++ Code/Web_app/client/src/App.tsx
@@ -13,9 +13,9 @@ import ProtectedRoute from "./routes/ProtectedRoute";
 import AdminHome from "./pages/admin/AdminHome";
 import FencerHome from "./pages/fencer/FencerHome";
 import TrainerHome from "./pages/trainer/TrainerHome";
-import CreateTrainer from "./pages/admin/CreateTrainer";
+import CreateTrainer from "./pages/trainer/CreateTrainer";
 import { Toolbar } from "@mui/material";
-import TrainerList from "./pages/admin/TrainerList";
+import TrainerList from "./pages/trainer/TrainerList";
 import TrainerProfile from "./pages/trainer/TrainerProfile";
 
 export const App = () => {
